reduce code triplication
